Renamed files to index.js in their respective folders
