+++ src/components/Project/index.js
@@ -1,8 +1,8 @@
 import React, { Component } from 'react';
 import { Container, Segment} from 'semantic-ui-react'
 
-import PackSmart from './pack_smart/pack_smart'
-import CarbonKnight from './carbon_knight/carbon_knight'
+import PackSmart from './pack_smart'
+import CarbonKnight from './carbon_knight'
 
 class Projects extends Component {
     render() {
